Change BlobIoVec.bi_size to u64
We were running into integer overflow issues with large (45GB) images with large (10GB) model weights. Changing bi_size to 64 bit fixed it.

+++ rafs/src/fs.rs
@@ -658,7 +658,7 @@ impl FileSystem for Rafs {
             let r = self.device.read_to(w, desc)?;
             result += r;
             recorder.mark_success(r);
-            if r as u32 != desc.size() {
+            if r as u64 != desc.size() {
                 break;
             }
         }
